bug modify

+++ ADFCommon/02.ADF.Entity/DictType.cs
@@ -8,15 +8,15 @@ namespace ADF.Entity
     public class DictType
     {
         [Key, Column(Order = 1)]
-        public Int32 MICDICT_ID { get; set; }
-        public Int16 TYPE_ID { get; set; }
-        public Int32 DICT_ID { get; set; }
+        public long MICDICT_ID { get; set; }
+        public int TYPE_ID { get; set; }
+        public int DICT_ID { get; set; }
         public String DICT_NAME { get; set; }
         public String NOTE { get; set; }
         public String PYM { get; set; }
         public String WBM { get; set; }
         public String HOSPITAL_NO { get; set; }
-        public Int16 STATE { get; set; }
+        public int STATE { get; set; }
 
     }
 }
